+++ vamp-test/src/pconley/vamp/util/test/PlaylistIteratorTest.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.NoSuchElementException;
 
-import pconley.vamp.model.Track;
+import pconley.vamp.library.model.Track;
 import pconley.vamp.util.Playlist;
 import pconley.vamp.util.PlaylistIterator;
 import android.net.Uri;
