+++ vamp-test/src/pconley/vamp/util/test/PlaylistTest.java
@@ -9,7 +9,7 @@ import java.util.Iterator;
 import java.util.LinkedList;
 import java.util.List;
 
-import pconley.vamp.model.Track;
+import pconley.vamp.library.model.Track;
 import pconley.vamp.util.Playlist;
 import android.net.Uri;
 import android.test.AndroidTestCase;
