+++ vamp/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:icon="@drawable/ic_launcher"
         android:label="@string/app_name" >
         <activity
-            android:name=".LibraryActivity"
+            android:name="pconley.vamp.library.LibraryActivity"
             android:label="@string/app_name" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -28,7 +28,7 @@
             android:name="pconley.vamp.player.PlayerActivity"
             android:label="@string/activity_player_name"
             android:launchMode="singleTop"
-            android:parentActivityName=".LibraryActivity" >
+            android:parentActivityName="pconley.vamp.library.LibraryActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value="pconley.vamp.LibraryActivity" />
