+++ vamp/res/layout/activity_library.xml
@@ -6,7 +6,7 @@
     android:paddingLeft="@dimen/activity_horizontal_margin"
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
-    tools:context="pconley.vamp.LibraryActivity" >
+    tools:context="pconley.vamp.library.LibraryActivity" >
 
     <ListView
         android:id="@+id/track_list"
