+++ vamp/src/pconley/vamp/player/PlayerService.java
@@ -4,8 +4,8 @@ import java.io.File;
 import java.io.IOException;
 
 import pconley.vamp.R;
-import pconley.vamp.db.TrackDAO;
-import pconley.vamp.model.Track;
+import pconley.vamp.library.db.TrackDAO;
+import pconley.vamp.library.model.Track;
 import pconley.vamp.util.BroadcastConstants;
 import pconley.vamp.util.Playlist;
 import pconley.vamp.util.PlaylistIterator;
