+++ vamp/src/pconley/vamp/util/Playlist.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 
-import pconley.vamp.model.Track;
+import pconley.vamp.library.model.Track;
 
 /**
  * A list of tracks. Built around a {@link List}, but with its {@link Iterator}
